Migrated away from Bevy for world data on the sever. It wasn't really bringing much. I was unable to drain the event queue while mutating the world resulting in server slowdown.

+++ renderer/renderer-ratatui/src/routes.rs
@@ -17,9 +17,9 @@ pub struct OrganizationView {
     pub selected_index: usize,
     pub organization_id: Uuid,
     pub tab: OrganizationTab,
-    pub marketing: u32,
-    pub rnd: u32,
-    pub training: u32,
+    pub marketing: u16,
+    pub rnd: u16,
+    pub training: u16,
 }
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
